size_t result for lengthProc. More usage of Tcl_GetWideUIntFromObj

+++ generic/tclArithSeries.c
@@ -462,7 +462,7 @@ TclArithSeriesObjIndex(Tcl_Obj *arithSeriesPtr, Tcl_WideInt index, Tcl_Obj **ele
  *
  *----------------------------------------------------------------------
  */
-unsigned long long TclArithSeriesObjLength(Tcl_Obj *arithSeriesPtr)
+size_t TclArithSeriesObjLength(Tcl_Obj *arithSeriesPtr)
 {
     ArithSeries *arithSeriesRepPtr = (ArithSeries*)
 	    arithSeriesPtr->internalRep.twoPtrValue.ptr1;
